+++ package.json
@@ -159,7 +159,6 @@
     "html-loader": "^1.1.0",
     "html-webpack-plugin": "^4.0.0-beta.2",
     "ignore-loader": "^0.1.2",
-    "immer": "^9.0.7",
     "jdenticon": "^2.1.1",
     "jquery": "^3.3.1",
     "jsdom": "^15.0.0",
